feat: 그래프 뷰를 담기 위한 GraphContainerView 파일 분리
